% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletdetection.R
\docType{class}
\name{doublet_detection}
\alias{doublet_detection}
\alias{normalize_counts}
\alias{load_10x_h5}
\alias{load_mtx}
\alias{load_10x_h5}
\alias{BoostClassifier-class}
\alias{BoostClassifier}
\title{Doublet detection in single-cell RNA-seq data}
\usage{
normalize_counts(raw_counts, pseudocount = 0.1)

load_10x_h5(file, genome = NULL, barcode_filtered = TRUE)

load_mtx(file)

load_10x_h5(file, genome = NULL, barcode_filtered = TRUE)
}
\arguments{
\item{raw_counts}{a matrix or data.frame containing raw UMI counts such as a gene-barcode matrix  (genes or transcripts by cell barcode).}

\item{pseudocount}{a numeric to add prior to log-transform (to avoid non-zero values).}

\item{file}{character: path to H5 file}

\item{genome}{character: path to top level H5 group}

\item{file}{character: path to mtx file}

\item{file}{character: path to H5 file}

\item{genome}{character: path to top level H5 group}

\item{raw_counts}{(array-like): Count matrix, oriented cells by genes.}
}
\description{
Functions required to compute identification of doublets in single-cell RNA-Seq experiments.
}
\section{Fields}{

\describe{
\item{\code{boost_rate}}{(numeric, optional): Proportion of cell population size to produce as synthetic doublets.}

\item{\code{n_components}}{(integer optional): Number of principal components used for clustering.}

\item{\code{n_top_var_genes}}{(integer optional): Number of highest variance genes to use; other genes discarded. Will use all genes when zero.}

\item{\code{new_lib_as:}}{(([integer integer]) -> integer optional): Method to use in choosing library size for synthetic doublets. Defaults to NULL which makes synthetic doublets the exact addition of its parents; alternative is new_lib_as = max.}

\item{\code{replace}}{(logical, optional): If FALSE, a cell will be selected as a synthetic doublet's parent no more than once.}

\item{\code{phenograph_parameters}}{(list, optional): Parameter dict to pass directly to Phenograph. Note that we change the Phenograph 'prune' default to TRUE; you must specifically include 'prune': FALSE here to change this.}

\item{\code{n_iters}}{(integer optional): Number of fit operations from which to collect p-values. Defualt value is 25. normalizer ((matrix) -> matrix): Method to normalize raw_counts. Defaults to normalize_counts, included in this package. Note: To use normalize_counts with its pseudocount parameter changed from the default 0.1 value to some positive numeric `new_var`, use: normalizer=lambda counts: doubletdetection.normalize_counts(counts, pseudocount=new_var)}

\item{\code{normalizer}}{((matrix) -> matrix): Method to normalize raw_counts. Defaults to normalize_counts, included in this package. Note: To use normalize_counts with its pseudocount parameter changed from the default 0.1 value to some positive numeric `new_var`, use: normalizer=lambda counts: doubletdetection.normalize_counts(counts, pseudocount=new_var)}

\item{\code{raw_counts}}{(array-like): Count matrix, oriented cells by genes.}

\item{\code{p_thresh}}{(numeric, optional): hypergeometric test p-value threshold that determines per iteration doublet calls}

\item{\code{voter_thresh}}{(numeric, optional): fraction of iterations a cell must be called a doublet}

\item{\code{cell1,cell2}}{(vector, numeric): Gene count vectors.}
}}


\keyword{doublets}
\keyword{filter}
\keyword{matrix}
\keyword{normalise}
\keyword{quality}
\keyword{scRNA}
\keyword{single-cell}
