% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletdetection.R
\name{load_10x_h5}
\alias{load_10x_h5}
\title{Load count matrix in 10x H5 format}
\usage{
load_10x_h5(file, genome = NULL, barcode_filtered = TRUE)
}
\arguments{
\item{file}{character: path to H5 file}

\item{genome}{character: path to top level H5 group}

\item{barcode_filtered}{logical: whether to use the filtered (TRUE) or raw (FALSE) gene-barcode matrix. Defaults to using filtered gene-barcode matrices.}
}
\description{
Functions required to compute identification of doublets in single-cell RNA-Seq experiments.

Adapted from: https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/advanced/h5_matrices
}
