% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletdetection.R
\name{load_mtx}
\alias{load_mtx}
\title{Load count matrix in mtx format}
\usage{
load_mtx(file)
}
\arguments{
\item{file}{character: path to mtx file}
}
\description{
Functions required to compute identification of doublets in single-cell RNA-Seq experiments.
}
