% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletdetection.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize count array.}
\usage{
normalize_counts(raw_counts, pseudocount = 0.1)
}
\arguments{
\item{raw_counts}{a matrix or data.frame containing raw UMI counts such as a gene-barcode matrix  (genes or transcripts by cell barcode).}

\item{pseudocount}{a numeric to add prior to log-transform (to avoid non-zero values).}
}
\description{
Functions required to compute identification of doublets in single-cell RNA-Seq experiments.
}
\keyword{doublets}
\keyword{filter}
\keyword{matrix}
\keyword{normalise}
\keyword{quality}
\keyword{scRNA}
\keyword{single-cell}
