% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubletdetection.R
\name{load_10x_h5}
\alias{load_10x_h5}
\title{Load count matrix in 10x H5 format}
\usage{
load_10x_h5(file, genome = NULL, barcode_filtered = TRUE)
}
\arguments{
\item{file}{character: path to H5 file}

\item{genome}{character: path to top level H5 group}
}
\description{
Adapted from: https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/advanced/h5_matrices
}
