% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{convergence}
\alias{convergence}
\alias{tsne_plot}
\title{Convergence plot of doublets}
\usage{
convergence(clf, show = FALSE, save = NULL, p_thresh = 0.99,
  voter_thresh = 0.9)

tsne_plot(raw_counts, labels, n_components = 30L, n_jobs = -1,
  show = False, save = None)
}
\arguments{
\item{clf}{(BoostClassifier object): Fitted classifier}

\item{show}{(logical, optional): If TRUE, runs plt.show()}

\item{save}{(charater, optional): filename for saved figure, figure not saved by default}

\item{p_thresh}{(numeric, optional): hypergeometric test p-value threshold that determines per iteration doublet calls}

\item{voter_thresh}{(numeric, optional): fraction of iterations a cell must be called a doublet}

\item{raw_counts}{(matrix): genes by cells count matrix}

\item{labels}{(vector, integer): predicted doublets from predict method}

\item{n_components}{(integer, optional): number of PCs to use prior to TSNE}

\item{n_jobs}{(integer, optional): number of cores to use for TSNE, -1 for all}

\item{show}{(logical, optional): If TRUE, runs plt.show()}

\item{save}{(charater, optional): filename for saved figure, figure not saved by default}
}
\description{
Produce a convergence plot showing number of cells called doublet per iteration

Produce a tsne plot of the data with doublets in black
}
\keyword{doublets}
\keyword{filter}
\keyword{matrix}
\keyword{normalise}
\keyword{quality}
\keyword{scRNA}
\keyword{single-cell}
